More patches (#13)
* Made sure the bot could deafen itself aswell as be able to find tracks

Bot has some problems deafining itself before and unable to find channel and message id's with the .split commands.

* Trying to fix .length issue

* Pushing stats to Top.gg

It's been awhile since I had this on Disconnect and have missed this. Finally re added it to fight the broken npm package I used to have posting stats.

* Mute certain servers from using the bot

This was as the request of the owner of the guild.